refactor: 패키지 구조 변경 및 네이밍 개선
- org.apache.coyote 패키지를 org.apache.catalina.*로 변경
- 패키지 역할에 맞게 클래스들의 위치 조정 및 가독성 향상
- 불필요한 import 수정 및 코드 정리
