+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -8,7 +8,10 @@ import java.io.InputStreamReader;
 import java.net.Socket;
 import java.net.URISyntaxException;
 
-import org.apache.coyote.*;
+import org.apache.catalina.controller.LoginController;
+import org.apache.catalina.controller.RegisterController;
+import org.apache.catalina.renderer.StaticRenderer;
+import org.apache.catalina.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
