+++ tomcat/src/main/java/org/apache/coyote/http11/Request.java
@@ -8,7 +8,7 @@ import java.io.IOException;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import org.apache.coyote.CookieManager;
+import org.apache.catalina.session.CookieManager;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
